# src/api/docs/swagger-old2.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "G-Taxi API",
    "description": "G-Taxi apis"
  },
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot Password",
        "description": "Forgot Password",
        "operationId": "forgot-password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgot_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forgot password in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset Password",
        "description": "Reset password",
        "operationId": "reset-password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/reset_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/practice/dashboard": {
      "get": {
        "tags": [
          "Practice"
        ],
        "summary": "dashboard",
        "description": "dashboard",
        "operationId": "dashboard",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "Show orders and categories listing",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
    "/practice/password": {
      "put": {
        "tags": [
          "Practice"
        ],
        "summary": "Change password",
        "description": "Change passord",
        "operationId": "password",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/change_password"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Change password in system",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/concierges/dashboard": {
      "get": {
        "tags": [
          "Concierges"
        ],
        "summary": "Dashboard for concierge type users",
        "description": "Dashboard for concierge type users",
        "operationId": "dashboard",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          },
          {
            "in": "query",
            "name": "searchValue"
          },
          {
            "in": "query",
            "name": "limit",
            "value": 10
          },
          {
            "in": "query",
            "name": "offset",
            "value": 0
          },
          {
            "in": "query",
            "name": "type",
            "value": "list"
          }
        ],
        "responses": {
          "200": {
            "description": "Get associated practices",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
    "/concierges/profile": {
      "get": {
        "tags": [
          "Concierges"
        ],
        "summary": "Get Concierge profile",
        "description": "Get Concierge profile",
        "operationId": "profile",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/HeaderVersion"
          },
          {
            "$ref": "#/parameters/HeaderLanguage"
          },
          {
            "$ref": "#/parameters/HeaderPlatform"
          },
          {
            "$ref": "#/parameters/HeaderAuth"
          }
        ],
        "responses": {
          "200": {
            "description": "Get concierge profile",
            "schema": {
              "$ref": "#/definitions/responsesOk"
            }
          },
          "500": {
            "description": "Unexpected error.",
            "schema": {
              "$ref": "#/definitions/responsesError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "reset_password": {
      "type": "object",
      "required": [
        "email",
        "password",
        "cnfpassword",
        "token",
        "otp"
      ],
      "properties": {
        "email": {
          "type": "string",
          "description": "Email",
          "example": "we@yopmail.com"
        },
        "password": {
          "type": "string",
          "format": "password",
          "example": "admin@1234"
        },
        "cnfpassword": {
          "type": "string",
          "format": "password",
          "example": "admin@1234"
        },
        "token": {
          "type": "string",
          "example": "dacsdvzdawre"
        },
        "otp": {
          "type": "string",
          "example": "1234"
        }
      }
    },
    "forgot_password": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "test@gmail.com"
        }
      }
    },
...
    "change_password": {
      "type": "object",
      "required": [
        "currentPassword",
        "newPassword",
        "cnfNewPassword"
      ],
      "properties": {
        "currentPassword": {
          "type": "string",
          "description": "Current Password",
          "format": "password",
          "example": "admin@1234"
        },
        "newPassword": {
          "type": "string",
          "description": "New Password",
          "format": "password",
          "example": "admin@12345"
        },
        "cnfNewPassword": {
          "type": "string",
          "description": "Confirm new Password",
          "format": "password",
          "example": "admin@12345"
        }
      }
    },
...
    "responsesOk": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "data": {
          "type": "object"
        }
      }
    },
    "responsesError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        },
        "error": {
          "type": "object"
        }
      }
    },
...
  },
  "parameters": {
    "HeaderVersion": {
      "name": "X-G-Taxi-Version",
      "in": "header",
      "description": "X-G-Taxi-Version",
      "required": true,
      "type": "string",
      "default": "1.0.0"
    },
    "HeaderLanguage": {
      "name": "Accept-Language",
      "in": "header",
      "description": "Accept-Language",
      "required": true,
      "type": "string",
      "default": "en"
    },
    "HeaderPlatform": {
      "name": "X-G-Taxi-Platform",
      "in": "header",
      "description": "android or ios or web",
      "required": true,
      "type": "string",
      "default": "android"
    },
    "HeaderAuth": {
      "name": "Authorization",
      "in": "header",
      "description": "Authorization",
      "required": true,
      "type": "string"
    },
...
  }
}